feature
feature

+++ backend/src/api/controllers/empresa.controller.js
@@ -1,5 +1,5 @@
-import prisma from "../../database/PrismaService";
-import usuarioService from "../services/usuario.service";
+import prisma from '../../database/PrismaService';
+import usuarioService from '../services/usuario.service';
 
 class EmpresaController {
   constructor() {}
@@ -19,17 +19,20 @@ class EmpresaController {
         empresa_id: empresa.id,
       };
 
-      const usuarioNovo =
-        await usuarioService.createUsuarioEmpresa(dadosUsuario);
+      const usuarioNovo = await usuarioService.createUsuarioEmpresa(dadosUsuario);
 
       reply.code(200).send({
         statusCode: 200,
-        message: "Empresa cadastrada com sucesso",
+        message: 'Empresa cadastrada com sucesso',
         data: { empresa, usuarioNovo },
       });
     } catch (error) {
       request.log.error(error);
-      reply.code(500).send(error);
+      reply.code(500).send({
+        statusCode: 500,
+        message: 'Ocorreu um erro no cadastro da empresa',
+        error: error.message,
+      });
     }
   }
 
@@ -45,10 +48,15 @@ class EmpresaController {
       reply.code(200).send({
         statusCode: 200,
         message: 'Empresa atualizada com sucesso',
+        data: empresa,
       });
-      } catch (error) {
+    } catch (error) {
       request.log.error(error);
-      reply.code(500).send(error);
+      reply.code(500).send({
+        statusCode: 500,
+        message: 'Ocorreu um erro na atualização da empresa',
+        error: error.message,
+      });
     }
   }
 
@@ -58,46 +66,75 @@ class EmpresaController {
       const empresa = await prisma.empresa.findFirst({ where: { id } });
 
       if (!empresa) {
-        reply.code(404).send({
+        return reply.code(404).send({
           statusCode: 404,
-          error: "Not Found",
-          message: "Empresa não encontrada",
+          error: 'Not Found',
+          message: 'Empresa não encontrada',
         });
       }
       reply.code(200).send(empresa);
     } catch (error) {
       request.log.error(error);
-      reply.code(500).send(error);
+      reply.code(500).send({
+        statusCode: 500,
+        message: 'Ocorreu um erro na busca da empresa',
+        error: error.message,
+      });
     }
   }
 
   async findMany(request, reply) {
     try {
       const empresas = await prisma.empresa.findMany({
-        orderBy: { id: "asc" },
+        orderBy: { id: 'asc' },
       });
 
       if (!empresas) {
-        reply.code(404).send({
+        return reply.code(404).send({
           statusCode: 404,
-          error: "Not Found",
-          message: "Nenhuma empresa encontrada",
+          error: 'Not Found',
+          message: 'Nenhuma empresa encontrada',
         });
       }
       reply.code(200).send(empresas);
     } catch (error) {
       request.log.error(error);
-      reply.code(500).send(error);
+      reply.code(500).send({
+        statusCode: 500,
+        message: 'Ocorreu um erro na busca das empresas',
+        error: error.message,
+      });
     }
   }
 
   async delete(request, reply) {
     try {
       const { id } = request.params;
+
+      const usuarioValidation = prisma.usuario.findFirst({ where: { empresa_id: id } });
+      const categoriaValidation = prisma.categoria.findFirst({ where: { empresa_id: id } });
+      const produtoValidation = prisma.produto.findFirst({ where: { empresa_id: id } });
+      const fornecedorValidation = prisma.fornecedor.findFirst({ where: { empresa_id: id } });
+      const controleProdutoValidation = prisma.controleProduto.findFirst({ where: { empresa_id: id } });
+
+      if (
+        usuarioValidation ||
+        categoriaValidation ||
+        produtoValidation ||
+        fornecedorValidation ||
+        controleProdutoValidation
+      ) {
+        return reply.code(400).send({
+          statusCode: 400,
+          error: 'Vínculado',
+          message: 'Esta empresa possui vínculos e não pode ser excluida',
+        });
+      }
+
       await prisma.empresa.delete({ where: { id } });
       reply.code(200).send({
         statusCode: 200,
-        message: "Empresa removida com sucesso",
+        message: 'Empresa removida com sucesso',
       });
     } catch (error) {
       request.log.error(error);
