+++ backend/src/api/controllers/perfil.controller.js
@@ -11,10 +11,15 @@ class PerfilController {
       reply.code(200).send({
         statusCode: 200,
         message: 'Novo perfil cadastrado com sucesso',
+        data: perfil,
       });
     } catch (error) {
       request.log.error(error);
-      reply.code(500).send(error);
+      reply.code(500).send({
+        statusCode: 500,
+        message: 'Ocorreu um erro no cadastro do perfil',
+        error: error.message,
+      });
     }
   }
 
@@ -30,10 +35,15 @@ class PerfilController {
       reply.code(200).send({
         statusCode: 200,
         message: 'Perfil atualizado com sucesso',
-      });    
+        data: perfil,
+      });
     } catch (error) {
       request.log.error(error);
-      reply.code(500).send(error);
+      reply.code(500).send({
+        statusCode: 500,
+        message: 'Ocorreu um erro na atualização do perfil',
+        error: error.message,
+      });
     }
   }
 
@@ -46,7 +56,7 @@ class PerfilController {
       });
 
       if (!perfil) {
-        reply.code(404).send({
+        return reply.code(404).send({
           statusCode: 404,
           error: 'Not Found',
           message: 'Nenhum perfil encontrado',
@@ -56,7 +66,11 @@ class PerfilController {
       reply.code(200).send(perfil);
     } catch (error) {
       request.log.error(error);
-      reply.code(500).send(error);
+      reply.code(500).send({
+        statusCode: 500,
+        message: 'Ocorreu um erro na busca do perfil',
+        error: error.message,
+      });
     }
   }
 
@@ -69,7 +83,7 @@ class PerfilController {
       });
 
       if (!perfis) {
-        reply.code(404).send({
+        return reply.code(404).send({
           statusCode: 404,
           error: 'Not Found',
           message: 'Nenhum perfil encontrado',
@@ -79,13 +93,27 @@ class PerfilController {
       reply.code(200).send(perfis);
     } catch (error) {
       request.log.error(error);
-      reply.code(500).send(error);
+      reply.code(500).send({
+        statusCode: 500,
+        message: 'Ocorreu um erro na busca dos perfis',
+        error: error.message,
+      });
     }
   }
 
   async delete(request, reply) {
     try {
       const { id } = request.params;
+
+      const perfilUsuarioValidation = await prisma.usuario.findFirst({ where: { perfil_id: id } });
+
+      if (perfilUsuarioValidation) {
+        return reply.code(400).send({
+          statusCode: 400,
+          error: 'Vínculado',
+          message: 'Este perfil está vínculado a um usuário',
+        });
+      }
 
       await prisma.perfil.delete({ where: { id } });
 
@@ -95,7 +123,11 @@ class PerfilController {
       });
     } catch (error) {
       request.log.error(error);
-      reply.code(500).send(error);
+      reply.code(500).send({
+        statusCode: 500,
+        message: 'Ocorreu um erro na exclusão do perfil',
+        error: error.message,
+      });
     }
   }
 }
